Fix axios warning and data tidy up
