+++ src/Assets/staticData.js
@@ -1,0 +1,48 @@
+import WrapPresentsImg from "./WrapPresents.jpg";
+import TidyGardenImg from "./TidyGarden.jpg";
+import ServeFoodImg from "./ServeFood.jpg";
+import SortClothingImg from "./SortClothing.jpg";
+
+import SweatShirt from "./SweatShirt.png"
+import Present from "./WrapPresents.png"
+import Food from "./ServeFood.png"
+import Garden from "./CleanGarden.png"
+
+export const taskImg = {
+  "Sort Clothes": SortClothingImg,
+  "Gardening": TidyGardenImg,
+  "Wrap Presents": WrapPresentsImg,
+  "Serve Food": ServeFoodImg,
+  "SweatShirt":SweatShirt,
+  "Present":Present,
+  "Food":Food,
+  "Garden":Garden,
+};
+
+export const waysToHelp = [
+    {
+      title: "Sort Clothing",
+      icon: "SweatShirt",
+      desc:
+        "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed ac odio eros. Suspendisse sollicitudin nisl quis erat venenatis, ut sodales nunc dignissim. Nullam ut risus ac nisi mollis vestibulum posuere nec leo. Mauris mattis risus quis lacinia condimentum. Morbi id sagittis erat. Vivamus ullamcorper nulla sit amet ullamcorper mattis. Sed maximus tellus eu nisl aliquet auctor vitae vehicula enim. Integer molestie nunc sed consequat finibus. Aenean posuere malesuada odio, et posuere mi gravida in.",
+    },
+    {
+      title: "Serve Food",
+      icon: "Food",
+      desc:
+        "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed ac odio eros. Suspendisse sollicitudin nisl quis erat venenatis, ut sodales nunc dignissim. Nullam ut risus ac nisi mollis vestibulum posuere nec leo. Mauris mattis risus quis lacinia condimentum. Morbi id sagittis erat. Vivamus ullamcorper nulla sit amet ullamcorper mattis. Sed maximus tellus eu nisl aliquet auctor vitae vehicula enim. Integer molestie nunc sed consequat finibus. Aenean posuere malesuada odio, et posuere mi gravida in.",
+    },
+    {
+      title: "Clean out a garden",
+      icon: "Garden",
+      desc:
+        "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed ac odio eros. Suspendisse sollicitudin nisl quis erat venenatis, ut sodales nunc dignissim. Nullam ut risus ac nisi mollis vestibulum posuere nec leo. Mauris mattis risus quis lacinia condimentum. Morbi id sagittis erat. Vivamus ullamcorper nulla sit amet ullamcorper mattis. Sed maximus tellus eu nisl aliquet auctor vitae vehicula enim. Integer molestie nunc sed consequat finibus. Aenean posuere malesuada odio, et posuere mi gravida in.",
+    },
+    {
+      title: "Wrap presents",
+      icon: "Present",
+      desc:
+        "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed ac odio eros. Suspendisse sollicitudin nisl quis erat venenatis, ut sodales nunc dignissim. Nullam ut risus ac nisi mollis vestibulum posuere nec leo. Mauris mattis risus quis lacinia condimentum. Morbi id sagittis erat. Vivamus ullamcorper nulla sit amet ullamcorper mattis. Sed maximus tellus eu nisl aliquet auctor vitae vehicula enim. Integer molestie nunc sed consequat finibus. Aenean posuere malesuada odio, et posuere mi gravida in.",
+    },
+    
+  ];
